+++ src/main/java/de/maxhenkel/delivery/integration/jei/NoDisplayTextComponent.java
@@ -1,0 +1,10 @@
+package de.maxhenkel.delivery.integration.jei;
+
+import net.minecraft.util.text.StringTextComponent;
+
+public class NoDisplayTextComponent extends StringTextComponent {
+
+    public NoDisplayTextComponent() {
+        super("");
+    }
+}